% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsZoop.R
\name{GibbsZoop}
\alias{GibbsZoop}
\title{Double Fusion-Calibration Zoop Models}
\usage{
GibbsZoop(
  Y1,
  Y2,
  T1,
  T2,
  X,
  date,
  coords,
  oneY1,
  oneY2,
  oneT1,
  oneT2,
  calibration = c("lm", "GP", "GPs", "cor", "corTime"),
  na = 0,
  nb = 1e-04,
  ga = 2,
  gb = 1,
  rangeTimesDmax = 3,
  sd = 1e-04,
  n.burnin = 10000,
  max.iter = 10000,
  n.thin = 10,
  n.report = 1000
)
}
\arguments{
\item{Y1}{log Surface zoop: \eqn{N_{Y1} \times 1}}

\item{Y2}{log Oblique zoop: \eqn{N_{Y2} \times 1}}

\item{T1}{CTD temperature: \eqn{N_{T1} \times 1}}

\item{T2}{therm temperature: \eqn{N_{T2} \times 1}}

\item{X}{covariates (whale, r2, r3): \eqn{N \times p}}

\item{date}{measurements date: \eqn{N \times 1}}

\item{coords}{measurements coordinate: \eqn{N \times 2}}

\item{oneY1}{binary wether \code{Y1} exists or not: \eqn{N \times 1}}

\item{oneY2}{binary wether \code{Y2} exists or not: \eqn{N \times 1}}

\item{oneT1}{binary wether \code{T1} exists or not: \eqn{N \times 1}}

\item{oneT2}{binary wether \code{T2} exists or not: \eqn{N \times 1}}

\item{calibration}{one of 
\code{c("lm", "GP", "GPs", "cor", "corTime")}
indicating a linear model (lm) in the calibration, a lm with
a spatially varying intercept, a lm with spatially varying 
intercept and slope, the same with coregionalization, and
the same with time varying coefs. in ZOOP calibration,
respectively}

\item{na, nb}{mean precision prior normal's}

\item{ga, gb}{shape rate prior gamma's}

\item{rangeTimesDmax}{the decay for the eta's is this value over dmax.
Suggestion, keep between 3 and 30.}

\item{sd}{tuned sd in rwMetropolis}

\item{n.burnin, max.iter, n.thin, n.report}{features MCMC}
}
\value{
Matrix with posterior samples from model parameters: 
  rows are iterations and columns are parameters
}
\description{
This function fits the double fusion and calibration zoop model
}
\author{
Jorge Castillo-Mateo
}
