% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictZoop.R
\name{predictZoop}
\alias{predictZoop}
\title{Zoop prediction}
\usage{
predictZoop(
  model,
  X,
  dateNew,
  coordsNew,
  dateData,
  coordsData,
  indY2 = NULL,
  indT2 = NULL,
  calibration = c("lm", "GP", "GPs", "cor", "corTime"),
  rangeTimesDmax = 3
)
}
\arguments{
\item{model}{model fitted with the function \code{\link{GibbsZoop}}}

\item{X}{covariates (whale, r2, r3): \eqn{1 \times p}}

\item{dateNew}{new date}

\item{coordsNew}{new coords}

\item{dateData}{dates used to fit the model}

\item{coordsData}{coords used to fit the model}

\item{indY2}{positions with Y2 not NA: \eqn{NY2 \times 1}}

\item{indT2}{positions with T2 not NA: \eqn{NT2 \times 1}}

\item{calibration}{one of 
\code{c("lm", "GP", "GPs", "cor", "corTime")}
indicating a linear model (lm) in the calibration, a lm with
a spatially varying intercept, a lm with spatially varying 
intercept and slope, the same with coregionalization, and
the same with time varying coefs. in ZOOP calibration,
respectively}

\item{rangeTimesDmax}{the same as used to fit the model}
}
\value{
matrix rows are simuations columns are:
  Y1 Y2 T1 T2 temp and log ZOOP
}
\description{
This function predicts Y1 Y2 T1 T2 temp and log ZOOP
}
\examples{
#a <- predict(model, 
#  X = c(1,0,0), 
#  dateNew = as.Date("2005-01-31"), coordsNew = c(-30000, 0),
#  dateData = date, coordsData = coords)

#b <- predict(model, 
#  X = c(1,0,0), 
#  dateNew = as.Date("2005-01-31"), coordsNew = c(-30000, 0),
#  dateData = date, coordsData = coords,
#  indY2, indT2, "GP")

}
