% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictZoopMap.R
\name{predictZoopMap}
\alias{predictZoopMap}
\title{ZOOP MAP Prediction}
\usage{
predictZoopMap(model, date = "2011-04-29", predictWhales, gamma0s, etats)
}
\arguments{
\item{model}{model fitted with the function \code{\link{GibbsZoop}}}

\item{date}{new date "yyyy-mm-dd"}

\item{predictWhales}{probabilities of whale presence, rows are simulations,
cols are days 181 x 17}

\item{gamma0s}{spatial field in temperature}

\item{etats}{spatial field in ZOOP. 
Choose the eta from the year to predict.}
}
\value{
matrix: rows are simulations columns are ZOOP
}
\description{
This function predicts ZOOP in the given day.
  The only covariates used are whale and regime (change if more are included).
  The spatial fields and probabilities of presence of whales must be 
  computed in advance and being introduced as arguments.
}
\author{
Jorge Castillo-Mateo
}
