% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{productNormals}
\alias{productNormals}
\title{COMPUTE PRODUCT OF NORMAL DENSITIES}
\usage{
productNormals(mu, prec)
}
\arguments{
\item{mu}{vector mean}

\item{prec}{vector precision}
}
\value{
vector(0) mean, vector(1) variance
}
\description{
COMPUTE PRODUCT OF NORMAL DENSITIES
}
