% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverZoop.R
\name{recoverZoop}
\alias{recoverZoop}
\title{Recover ZOOP}
\usage{
recoverZoop(model, X, date, coords, do.mean = FALSE, log.scale = FALSE)
}
\arguments{
\item{model}{model fitted with the function \code{\link{GibbsZoop}}}

\item{X}{covariates (whale, r2, r3?): \eqn{N \times p}}

\item{date}{\eqn{N \times 1}}

\item{coords}{\eqn{N \times 2}}

\item{do.mean}{return the mean? TRUE Or samples? FALSE}

\item{log.scale}{return log-scale? TRUE Or original scale? FALSE}
}
\value{
matrix: rows are simuations columns are (log) ZOOP
}
\description{
This function recovers ZOOP from a fitted model
}
\examples{
#ZOOP <- recoverZoop(model, X, date, coords)
#plot(x=ZOOP, y=auxY1)
#plot(x=ZOOP, y=auxY2)
# 
#beta <- matrix(nrow = 17, ncol = 4)
#for (i in 1:17) {
#   beta[i,] <- coefficients(lm(auxY2 ~ ZOOP, subset = year(date) == i + 2002))
#}
#plot(beta[,1])
#plot(beta[,2])
#plot(beta[,3])
#plot(beta[,4])

}
