% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Validation RMSE MAE CRPS CVG}
\usage{
validation(Y, Yhat, alpha = 0.1)
}
\arguments{
\item{Y}{actual data}

\item{Yhat}{replicated or simulated data, rows are replicates 
columns are data}

\item{alpha}{CVG level}
}
\value{
matrix with RMSE, MAE, CRPS and CVG
}
\description{
This function validates the model
}
