% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictZoop.R
\name{predictZoop}
\alias{predictZoop}
\title{Predict Function for Zoop Models}
\usage{
predictZoop(
  model,
  X,
  dateNew,
  coordsNew,
  dateData,
  coordsData,
  indY2 = NULL,
  indT2 = NULL,
  calibration = c("lm", "GP", "GPs", "cor", "corTime"),
  rangeTimesDmax = 3
)
}
\arguments{
\item{model}{model fitted with the function \code{\link{GibbsZoop}}}

\item{X}{covariates (whale, r2, r3): \eqn{1 \times p}}

\item{dateNew}{new date}

\item{coordsNew}{new coords}

\item{dateData}{dates used to fit the model}

\item{coordsData}{coords used to fit the model}

\item{indY2}{positions with \code{Y2} not NA: \eqn{NY2 \times 1}}

\item{indT2}{positions with \code{T2} not NA: \eqn{NT2 \times 1}}

\item{calibration}{same as in \code{\link{GibbsZoop}}}

\item{rangeTimesDmax}{same as in \code{\link{GibbsZoop}}}
}
\value{
matrix rows are simuations columns are:
  Y1 Y2 T1 T2 temp and log ZOOP
}
\description{
This function predicts \code{Y1}, \code{Y2}, \code{T1}, 
  \code{T2}, true temp, and true log ZOOP based on \code{\link{GibbsZoop}}
  models
}
\examples{
#a <- predict(model, 
#  X = c(1,0,0), 
#  dateNew = as.Date("2005-01-31"), coordsNew = c(-30000, 0),
#  dateData = date, coordsData = coords)

#b <- predict(model, 
#  X = c(1,0,0), 
#  dateNew = as.Date("2005-01-31"), coordsNew = c(-30000, 0),
#  dateData = date, coordsData = coords,
#  indY2, indT2, "GP")

}
\author{
Jorge Castillo-Mateo
}
